% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_lbls.R
\name{age_lbls}
\alias{age_lbls}
\title{Clean up age group labels}
\usage{
age_lbls(x)
}
\arguments{
\item{x}{A string vector of labels to clean up}
}
\value{
A string vector of display-worthy labels
}
\description{
This function does some tedious regex replacement to make common labels we use for age groups (i.e. syntactically correct variable names) to labels meant for display in charts and tables. The main features are replacing underscores between numbers with dashes, adding space between text and numbers, replacing symbols, and parsing numbers. So "ages00_17" becomes "Ages 0-17", and "ages65plus" becomes "Ages 65+". Of course, you could also use this for other numeric ranges, like years.
}
\examples{
a <- c("ages00_17", "ages18_64", "ages65plus", "under18", "ages18up")
age_lbls(a)
}
\keyword{string-formatting}
