% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cws2018}
\alias{cws2018}
\alias{chronic_disease}
\alias{community_cohesion}
\alias{fin_insecurity}
\alias{walkability}
\title{2018 DCWS sample indicators}
\format{
Data frames with 4 columns:
\describe{
\item{question}{Abbreviated indicator}
\item{category}{Category: gender, age, etc.}
\item{group}{Group: male, female, ages 18–34, etc.}
\item{value}{Percentage value}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 69 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 69 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 69 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46 rows and 4 columns.
}
\source{
2018 DataHaven Community Wellbeing Survey
}
\usage{
chronic_disease

community_cohesion

fin_insecurity

walkability
}
\description{
This is a subset of the 2018 DataHaven Community Wellbeing Survey data for Greater New Haven covering several commonly used indicators, split into separate data frames for each topic but with the same format.
}
\details{
Detail and meaning of values for each dataset and question:\tabular{lll}{
   Dataset \tab Question \tab Definition / response \cr
   chronic_disease \tab hypertension \tab Received hypertension diagnosis \cr
   chronic_disease \tab asthma \tab Received asthma diagnosis \cr
   chronic_disease \tab diabetes \tab Received diabetes diagnosis \cr
   community_cohesion \tab positive_role_models \tab Strongly / somewhat agree local youth have positive role models \cr
   community_cohesion \tab satisfied_with_area \tab Satisfied with area \cr
   community_cohesion \tab trust_neighbors \tab Strongly / somewhat agree neighbors can be trusted \cr
   fin_insecurity \tab food_insecurity \tab Unable to afford food in past 12 months \cr
   fin_insecurity \tab housing_insecurity \tab Unable to afford adequate housing in past 12 months \cr
   fin_insecurity \tab no_bank_account \tab Didn't have bank account in past 12 months \cr
   walkability \tab safe_biking \tab Strongly / somewhat agree there are safe places to bike nearby \cr
   walkability \tab safe_sidewalks \tab Strongly / somewhat agree there are safe sidewalks & crosswalks nearby \cr
}
}
\keyword{datasets}
