% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_exp}
\alias{life_exp}
\title{Life expectancy}
\format{
A data frame with 189 rows and 3 variables:
\describe{
\item{tract}{Tract FIPS code}
\item{town}{Town name}
\item{value}{Life expectancy, including \code{NA} values for 9 tracts without data}
}
}
\source{
\href{https://www.cdc.gov/nchs/nvss/usaleep/usaleep.html}{CDC's USALEEP project}, current as of Oct. 2021
}
\usage{
life_exp
}
\description{
Dataset of life expectancy in years by census tract within New Haven County for 2010 to 2015.
}
\keyword{datasets}
