% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{median_age}
\alias{median_age}
\title{Median age by sex}
\format{
A data frame with 141 rows and 5 variables:
\describe{
\item{level}{Geographic level: factor with levels for state, county, and town}
\item{county}{County: \code{NA} for state- and county- level observations, county name for town-level observations}
\item{name}{Name of geography}
\item{sex}{Sex: factor with levels total, male, and female}
\item{value}{Median age}
}
}
\source{
2023 US Census Bureau American Community Survey 5-year estimates
}
\usage{
median_age
}
\description{
This is a sample of data from the 2023 ACS, table B01002. Values are median age for Connecticut, the Capitol Region and Greater Bridgeport councils of government (COGs), and all towns in either COG, each for total population, males, and females.
}
\examples{
head(median_age)

}
\keyword{datasets}
