% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_formats.R
\name{percent100}
\alias{percent100}
\alias{percent_lt1}
\alias{round100}
\title{Number formatting utilities}
\usage{
percent100(x, ...)

percent_lt1(x, ...)

round100(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{Additional arguments to pass to underlying \code{scales} functions, such as \code{scales::percent} and \code{scales::number}.}
}
\value{
A string vector of formatted numbers.
}
\description{
This is a collection of small utilities for formatting numbers, especially for use in charts and tables. Many of them are simply wrappers around \code{scales} functions.
}
\details{
\code{round100} takes a decimal, such as a percentage, multiplies by 100, and rounds, e.g. \code{0.251} becomes \code{"25"}, useful for then pasting with additional text (see examples).

\code{percent100} adds on to \code{round100} by appending a percent sign, e.g. \code{0.251} becomes \code{"25\%"}.

\code{percent_lt1} takes an additional step, replacing any values below 1 percent with \code{"<1\%"}, useful for suppressing small numbers.
}
\examples{
paste("The value is", round100(0.251), "percent.") # outputs "The value is 25 percent."

percent100(c(0.25, 0.251, 0.008)) # outputs "25\%", "25\%", "1\%"

percent_lt1(c(0.25, 0.251, 0.008)) # outputs "25\%", "25\%", "<1\%"
}
