% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_thresh.R
\name{number_thresh}
\alias{number_thresh}
\alias{percent_thresh}
\alias{dollar_thresh}
\title{Format and top- or bottom-code numbers based on a threshold}
\usage{
number_thresh(
  x,
  thresh,
  less_than = TRUE,
  accuracy = 1,
  txt = FALSE,
  html = FALSE,
  ...
)

percent_thresh(
  x,
  thresh,
  less_than = TRUE,
  accuracy = 1,
  txt = FALSE,
  html = FALSE,
  ...
)

dollar_thresh(
  x,
  thresh,
  less_than = TRUE,
  accuracy = 1,
  txt = FALSE,
  html = FALSE,
  ...
)
}
\arguments{
\item{x}{Numeric vector}

\item{thresh}{Numeric, the threshold above/below which numbers will be capped.}

\item{less_than}{Boolean: if \code{TRUE}, values \emph{less than} the threshold will be
lumped together. Otherwise, values \emph{greater than} the threshold will be
lumped. Ignored if both bottom and top endpoints are given in \code{thresh}. Default: TRUE}

\item{accuracy}{Number: accuracy of formatted numbers, passed to \code{scales::label_number}
and related functions. Defaults to 1, meaning no decimal places are returned.}

\item{txt}{Boolean: if \code{TRUE}, plain English is used (e.g. "less than") instead of
symbols (e.g. "<"). For \code{percent_thresh}, this also means using " percent" instead
of "\%". Default: FALSE}

\item{html}{Boolean: if \code{TRUE}, HTML-appropriate symbols are used (e.g. \verb{&lt;})
instead of more readable ones (e.g. "<"). Has no effect if \code{txt = TRUE}. Default: FALSE
function}

\item{...}{
  Arguments passed on to \code{\link[scales:label_number]{scales::label_number}}
  \describe{
    \item{\code{scale}}{A scaling factor: \code{x} will be multiplied by \code{scale} before
formatting. This is useful if the underlying data is very small or very
large.}
    \item{\code{prefix}}{Additional text to display before the number. The suffix is
applied to absolute value before \code{style_positive} and \code{style_negative} are
processed so that \code{prefix = "$"} will yield (e.g.) \verb{-$1} and \verb{($1)}.}
    \item{\code{suffix}}{Additional text to display after the number.}
    \item{\code{big.mark}}{Character used between every 3 digits to separate thousands.}
    \item{\code{decimal.mark}}{The character to be used to indicate the numeric
decimal point.}
    \item{\code{style_positive}}{A string that determines the style of positive numbers:
\itemize{
\item \code{"none"} (the default): no change, e.g. \code{1}.
\item \code{"plus"}: preceded by \code{+}, e.g. \code{+1}.
\item \code{"space"}: preceded by a Unicode "figure space", i.e., a space equally
as wide as a number or \code{+}. Compared to \code{"none"}, adding a figure space
can ensure numbers remain properly aligned when they are left- or
right-justified.
}}
    \item{\code{style_negative}}{A string that determines the style of negative numbers:
\itemize{
\item \code{"hyphen"} (the default): preceded by a standard hypen \code{-}, e.g. \code{-1}.
\item \code{"minus"}, uses a proper Unicode minus symbol. This is a typographical
nicety that ensures \code{-} aligns with the horizontal bar of the
the horizontal bar of \code{+}.
\item \code{"parens"}, wrapped in parentheses, e.g. \code{(1)}.
}}
    \item{\code{scale_cut}}{Named numeric vector that allows you to rescale large
(or small) numbers and add a prefix. Built-in helpers include:
\itemize{
\item \code{cut_short_scale()}: [10^3, 10^6) = K, [10^6, 10^9) = M, [10^9, 10^12) = B, [10^12, Inf) = T.
\item \code{cut_long_scale()}: [10^3, 10^6) = K, [10^6, 10^12) = M, [10^12, 10^18) = B, [10^18, Inf) = T.
\item \code{cut_si(unit)}: uses standard SI units.
}

If you supply a vector \code{c(a = 100, b = 1000)}, absolute values in the
range \verb{[0, 100)} will not be rescaled, absolute values in the range \verb{[100, 1000)}
will be divided by 100 and given the suffix "a", and absolute values in
the range \verb{[1000, Inf)} will be divided by 1000 and given the suffix "b".
If the division creates an irrational value (or one with many digits), the
cut value below will be tried to see if it improves the look of the final
label.}
    \item{\code{trim}}{Logical, if \code{FALSE}, values are right-justified to a common
width (see \code{\link[base:format]{base::format()}}).}
  }}
}
\value{
A character vector of the same length as \code{x}
}
\description{
This is a set of helper functions for formatting numbers where you
also need to top-code or bottom-code values above or below some threshold (or both).
Additional arguments help with generating plain English text or escaped HTML
characters for printing in tables.
}
\examples{
rate <- c(0.9, 0.95, 0.92, 0.991)
percent_thresh(rate, thresh = 0.99, less_than = FALSE)
percent_thresh(rate, thresh = 0.99, less_than = FALSE, txt = TRUE)

# censor amounts under 100 dollars or above 1000 dollars
money <- c(200, 99, 400, 1005, 999)
dollar_thresh(money, thresh = c(100, 1000))
}
\keyword{string-formatting}
