% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rent_x_income}
\alias{rent_x_income}
\title{Rent and income data}
\format{
A data frame with 48 rows and 6 variables:
\describe{
\item{name}{Town name}
\item{bedrooms}{Whether values correspond to all housing units, or just 2-bedroom units}
\item{renters}{Number of renter households}
\item{income}{Median household income of renter households}
\item{annual_rent}{Annual rent, calculated as 12 * median gross rent}
\item{income_needed}{Income needed for the median rent to make up no more than 30 percent of a household's income, per HUD housing affordability guidelines. Beyond this, a household is considered cost-burdened.}
}
}
\source{
US Census Bureau 2019 American Community Survey tables B25119, B25031, and B25042.
}
\usage{
rent_x_income
}
\description{
This is data from a few tables of the 2019 American Community Survey for New Haven County, useful for comparing household incomes to local rents as a way to measure rent affordability.
}
\keyword{datasets}
