% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_barcontinuous.R
\name{scale_y_barcontinuous}
\alias{scale_y_barcontinuous}
\alias{scale_x_barcontinuous}
\title{Set a continuous scale more suited to bar charts}
\usage{
scale_y_barcontinuous(top = 0.05, ...)

scale_x_barcontinuous(top = 0.05, ...)
}
\arguments{
\item{top}{Numeric, giving the ratio by which to set the upper padding. Defaults 0.05, as is the case for the underlying \code{ggplot2::scale_x/y_continuous}.}

\item{...}{Additional arguments passed on to \code{ggplot2::scale_x/y_continuous}.}
}
\description{
This is a light wrapper around \code{ggplot2::scale_x_continuous} / \code{ggplot2::scale_y_continuous} that sets an \code{expand} argument of \code{mult = c(0, 0.05)}. The purpose of this is just to align the bases of bars with the x-axis (for \code{scale_y_continuous}) or y-axis (for \code{scale_x_continuous}), rather than the default that leaves a gap between bars and the x- or y-axis.
}
\keyword{ggplot}
\keyword{viz-utils}
