% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{self_rated_health}
\alias{self_rated_health}
\title{2018 DCWS sample indicators: self-rated health}
\format{
A data frame with 100 rows and 4 variables:
\describe{
\item{category}{Category: gender, age, etc.}
\item{group}{Group: male, female, ages 18–34, etc.}
\item{response}{Response: excellent, very good, good, fair, poor}
\item{value}{Percentage value}
}
}
\source{
2018 DataHaven Community Wellbeing Survey
}
\usage{
self_rated_health
}
\description{
This is a subset of the 2018 DataHaven Community Wellbeing Survey data for Greater New Haven, just including the question on self-rated health. Unlike \link[=cws2018]{other DCWS datasets} in this package, this includes all response options.
}
\keyword{datasets}
