% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{title_case}
\alias{title_case}
\title{String utilities}
\usage{
title_case(x, lower = NULL)
}
\arguments{
\item{x}{A character vector to alter}

\item{lower}{Aside from a built-in list of articles and prepositions, any additional words that should be kept lowercase.}
}
\value{
A character vector
}
\description{
This is a collection of a few functions for common string transformations.
}
\details{
\code{title_case} is a wrapper around \code{stringr::str_to_title}, which corrects for the fact that not \emph{all} words in a title should be capitalized (articles, prepositions, etc. not at the beginning of the title). It follows the APA style guide the best I could, but there may be exceptions to fix manually (see examples).
}
\examples{
title_case(c("Rates are rising in the area", "Of all adults in the region"))
title_case("Rates are going up in the area") # This result is technically
# incorrect--"up" is used as an adverb here, and under APA should be
# capitalized. Write this one out manually.
}
