% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cws_trend}
\alias{cws_trend}
\title{2015–2018 DCWS sample trends}
\format{
A data frame with 48 rows and 5 variables:
\describe{
\item{year}{Year, either 2015 or 2021}
\item{indicator}{Indicator: satisfied with area, local government is responsive, or police approval}
\item{category}{Category: gender, age, etc.}
\item{group}{Group: male, female, ages 18–34, etc.}
\item{value}{Percentage value}
}
}
\source{
2015 & 2021 DataHaven Community Wellbeing Survey
}
\usage{
cws_trend
}
\description{
This is a subset of 3 questions from both the 2015 & 2021 DataHaven Community Wellbeing Surveys for Connecticut.
}
\keyword{datasets}
