% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodge_lbls.R
\name{dodge_lbls}
\alias{dodge_lbls}
\title{Prepare chart labels that should be repelled apart}
\usage{
dodge_lbls(data, x, value, group, thresh, digits = 2, verbose = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{x}{Bare name of the column being plotted on the independent axis}

\item{value}{Bare name of the dependent variable}

\item{group}{Bare column name of the grouping variable, likely used for
point color}

\item{thresh}{Numeric: the threshold value of the difference between points.
Any pair of points with a difference of less than \code{thresh} will be
considered too close, and will be included in the output as points to dodge.}

\item{digits}{Number of digits to round to before calculating differences
between points. Default: 2}

\item{verbose}{Boolean: if \code{TRUE}, will report the number of observations being dodged. Defaults \code{FALSE}.}
}
\value{
A data frame of too-close points, with columns corresponding to the
\code{x} and \code{group} columns.
}
\description{
This function is designed for plots, particularly dot plots,
where some values are very close together and need to be pushed apart in
order to avoid overlaps. It returns just the columns needed to know which
points should be dodged apart.

It doesn't calculate new positions for those points—
that depends on your specific use, but \code{ggrepel::geom_text_repel} is likely
what you'll use with this. You'll probably call this to get a data frame
of the points that should be dodged, then use \code{dplyr::semi_join} on the
original data to plot the points that \emph{do} need to be dodged, and
\code{dplyr::anti_join} for the ones that \emph{don't} need to be dodged (see example).
}
\examples{
if (requireNamespace("ggrepel", quietly = TRUE)) {
  library(ggplot2)
  
  cws24 <- cws_trend |>
    dplyr::filter(year == 2024)
  
  # some labels are overlapped & hard to read
  ggplot(cws24, aes(x = value, y = question)) +
    geom_point(aes(color = group), size = 6, alpha = 0.8) +
    geom_text(aes(label = percent100(value)), size = 3)
  
  to_dodge <- dodge_lbls(cws24, 
                         x = question, value = value, group = group, 
                         thresh = 0.01)
  
  # the first text geom will be observations that don't need to be dodged (anti-join)
  # the second will be ones that do need to be dodged (semi-join)
  ggplot(cws24, aes(x = value, y = question)) +
    geom_point(aes(color = group), size = 6, alpha = 0.8) +
    geom_text(aes(label = percent100(value)), size = 2.7,
              data = ~dplyr::anti_join(., to_dodge, by = c("question", "group"))) +
    ggrepel::geom_text_repel(aes(label = percent100(value), color = group), 
                             size = 2.7, seed = 1,
                             data = ~dplyr::semi_join(., to_dodge, by = c("question", "group")),
                             direction = "x", nudge_y = 0.2)
}
}
\keyword{plot-labels}
\keyword{viz-utils}
