% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_lbls.R
\name{endpoint_lbls}
\alias{endpoint_lbls}
\title{Formatted labels for endpoints of trend charts}
\usage{
endpoint_lbls(
  data,
  x,
  value,
  group,
  mult = 0.05,
  add = NULL,
  fun = NULL,
  long_side = c("right", "left", "both", "none")
)
}
\arguments{
\item{data}{A data frame}

\item{x}{Bare column name of the independent value which has the start and finish points (probably a date). If this isn't numeric, \code{endpoint_lbls} will try to coerce it to numeric, which may yield undesired results.}

\item{value}{Bare column name of the value column}

\item{group}{Bare column name of the name or grouping variable that should be shown in the labels.}

\item{mult}{Numeric; ratio to the \code{x} variable by which labels should be offset. Default: 0.05.}

\item{add}{Numeric; single value along the \code{x} variable by which labels should be offset. Defaults \code{NULL}. If both \code{mult} and \code{add} are given, \code{mult} takes precedence. Previously there was a \code{frac} argument which was supposed to act like \code{mult} but erroneously acted like \code{add}.}

\item{fun}{A function, used to create value labels. \verb{scales::label_*} functions will be very useful here. If \code{NULL} (the default), no formatting is done.}

\item{long_side}{Character, either \code{"right"} (the default), \code{"left"}, \code{"both"}, or \code{"none"}. For \code{"left"} or \code{"right"}, this refers to whether the longer label should be on the right or the left, returning a short label on the opposite side. If \code{"both"}, only long labels are returned; if \code{"none"}, only short labels are returned. Regardless, labels are combined in a new column called \code{lbl}.}
}
\value{
A data frame: the original data frame passed in to \code{data}, with 3 additional columns:
\itemize{
\item \code{x}, the x-values with offsets added
\item \code{just}, a column of 0 or 1 giving the justification value dependent on which end the label will appear
\item \code{lbl}, a column of formatted label text
}
}
\description{
Easily put together labels for the endpoints of a chart, such as a line chart with outwardly-justified labels at each end. Labels with the formats \code{{name}: {fun(value)}} (long format) and \verb{\{fun(value\}} (short format) will be created, such as \code{"Connecticut: 12\%"} and \code{"16\%"} respectively. Often one format will be used on one end and the other format on the other end, but \code{long_side} gives you some options around this. The intention around combining labels like this is so you can put labels on both ends, aligned nicely, in a single \code{geom_text} call.
}
\examples{
library(ggplot2)
# note that it will still be up to your judgment to set scale expansion,
# since that will depend on things outside the scope of just this function
cws_trend |>
    dplyr::filter(question == "local_govt_responsive", category == "Age") |>
    endpoint_lbls(
        value = value, x = year, group = group,
        fun = percent100
    ) |>
    ggplot(aes(x = year, y = value, color = group)) +
    geom_line() +
    geom_point(size = 3) +
    geom_text(aes(label = lbl, hjust = just, x = x)) +
    scale_x_continuous(
        expand = expansion(add = c(1, 3)),
        breaks = c(2015, 2024)
    ) +
    theme(legend.position = "none")

cws_trend |>
    dplyr::filter(question == "local_govt_responsive", category == "Age") |>
    endpoint_lbls(
        value = value, x = year, group = group, long_side = "both",
        fun = percent100, add = 0.4, mult = NULL
    ) |>
    ggplot(aes(x = year, y = value, color = group)) +
    geom_line() +
    geom_point(size = 3) +
    geom_text(aes(label = lbl, hjust = just, x = x)) +
    scale_x_continuous(
        expand = expansion(add = 3.5),
        breaks = c(2015, 2024)
    ) +
    theme(legend.position = "none")
}
\keyword{plot-labels}
\keyword{viz-utils}
