% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_here.R
\name{glue_here}
\alias{glue_here}
\title{Parameterized, project-based file paths}
\usage{
glue_here(x, ..., .snake = TRUE)
}
\arguments{
\item{x}{Character vector of paths, relative to the project root, with glue-style encodings.}

\item{...}{Additional named arguments to fill into glue brackets}

\item{.snake}{Boolean, whether text should also be converted to snakecase. Default: TRUE}
}
\value{
A character vector of absolute paths with glue strings filled in.
}
\description{
Quick wrapper around \code{stringr::str_glue} and \code{here::here} to fill a small recurring need: filling in file paths based on some parameter, but building them relative to the project root.
}
\examples{
year <- 2025
loc <- "New Haven"
glue_here(c("input_data/demo/{loc}_table_{year}.csv",
            "{year}_headings.txt"))
# --> projectdir/input_data/demo/new_haven_table_2025.csv,
#       projectdir/2025_headings.txt
glue_here("{year}_headings_{topic}.txt", topic = "education")
# --> projectdir/2025_headings_education.txt
}
\seealso{
\code{\link[here:here]{here::here()}} \code{\link[stringr:str_glue]{stringr::str_glue()}}
}
