% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_formats.R
\name{round100}
\alias{round100}
\alias{comma}
\alias{percent100}
\alias{percent_lt1}
\alias{percent_txt}
\alias{percent_eng}
\alias{dollark}
\alias{dollar1}
\title{Number formatting utilities}
\usage{
round100(x, ...)

comma(x, ...)

percent100(x, ...)

percent_lt1(x, ...)

percent_txt(x, accuracy = 1, ...)

percent_eng(x, accuracy = 1, sentence_case = TRUE)

dollark(x, accuracy = 1, ...)

dollar1(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{Additional arguments to pass to underlying \code{scales} functions, such as \code{scales::label_percent} and \code{scales::label_number}.}

\item{accuracy}{A number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}

\item{sentence_case}{Boolean: if \code{TRUE} (default), text will also be converted to sentence case,
i.e. capitalizing the first letter}
}
\value{
A string vector of formatted numbers.
}
\description{
This is a collection of small utilities for formatting numbers, especially for use in charts and tables. Many of them are simply wrappers around \code{scales} functions.
}
\details{
\code{round100} takes a decimal, such as a percentage, multiplies by 100, and rounds, e.g. \code{0.251} becomes \code{"25"}, useful for then pasting with additional text (see examples).

\code{percent100} adds on to \code{round100} by appending a percent sign, e.g. \code{0.251} becomes \code{"25\%"}.

\code{percent_lt1} takes an additional step, replacing any values below 1 percent with \code{"<1\%"}, useful for suppressing small numbers.
}
\examples{
paste("The value is", percent_txt(0.25))

percent100(c(0.25, 0.251, 0.008))

percent_lt1(c(0.25, 0.251, 0.008))

paste(percent_eng(0.1), "of adults say...")

dollar1(c(81.1, 105.22))

dollark(c(12345, 89100))
dollark(c(12345, 89100), accuracy = 0.1)
}
\keyword{string-formatting}
