% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_case.R
\name{title_case}
\alias{title_case}
\title{Title casing}
\usage{
title_case(x, lower = NULL, clean = TRUE, ...)
}
\arguments{
\item{x}{A character vector to alter}

\item{lower}{Aside from a built-in list of articles and prepositions, any additional words that should be kept lowercase.}

\item{clean}{Boolean: if \code{TRUE} (the default), strings will be passed to \code{clean_titles} as an intermediate step, mostly to break snakecased text into words before converting to title casing.}

\item{...}{
  Arguments passed on to \code{\link[=clean_titles]{clean_titles}}
  \describe{
    \item{\code{cap_all}}{Logical: if \code{TRUE}, first letter of each word after splitting will be capitalized. If \code{FALSE}, only the first character of the string will be capitalized. Note that in order to balance this with respecting consecutive capital letters, such as from acronyms,}
    \item{\code{split_case}}{Logical: if \code{TRUE}, consecutive lowercase-uppercase pairs will be treated as two words to be separated.}
    \item{\code{keep_running_caps}}{Logical: if \code{TRUE}, consecutive uppercase letters will be kept uppercase.}
    \item{\code{space}}{Character vector of characters and/or regex patterns that should be replaced with a space to separate words.}
    \item{\code{remove}}{Character vector of characters and/or regex patterns that will be removed before any other operations; if \code{NULL}, nothing is removed.}
  }}
}
\value{
A character vector
}
\description{
This function is a wrapper around \code{stringr} functions, which corrects for the fact that not \emph{all} words in a title should be capitalized (articles, prepositions, etc. not at the beginning of the title). It follows the APA style guide the best I could, but there may be exceptions to fix manually (see examples).
}
\examples{
title_case(c("rates are rising in the area", "of all adults in the region"))
title_case("rates are going up in the area") 
# This result is technically
# incorrect--"up" is used as an adverb here, and under APA should be
# capitalized. Write this one out manually.
}
\seealso{
\code{\link[stringr:case]{stringr::str_to_title()}}, \code{\link[=clean_titles]{clean_titles()}}
}
\keyword{string-formatting}
